Fix broken compilation on linux
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3196 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ apps/samsonPop/main_samsonPop.cpp
@@ -176,7 +176,7 @@ int main( int argC , const char *argV[] )
             delete block;
             
             if( max_size > 0 )
-                if( downloaded_content >= max_size )
+			   if( downloaded_content >= (size_t)max_size )
                     break;
 
         }
